login and registration is now in seperate controller
