+++ GithubJobsEnterpriseProject/Controllers/GithubJobsController.cs
@@ -2,33 +2,28 @@
 using GithubJobsEnterpriseProject.Services;
 using Microsoft.AspNetCore.Authentication;
 using Microsoft.AspNetCore.Authentication.Cookies;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Net;
 using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace GithubJobsEnterpriseProject.Controllers
 {
-    [Route("api")]
+    [Route("/api")]
     [ApiController]
     public class GithubJobsController : ControllerBase
     {
         private readonly JobContext _context;
-        private HashService _hashService = new HashService();
-        private readonly UserContext _userContext;
         private readonly IJobApiService _apiService;
-             
-        public GithubJobsController(JobContext context,UserContext userContext, IJobApiService apiService)
+
+        public GithubJobsController(JobContext context, UserContext userContext, IJobApiService apiService)
         {
 
             _context = context;
-            _userContext = userContext;
             _apiService = apiService;
         }
 
@@ -155,74 +150,5 @@ namespace GithubJobsEnterpriseProject.Controllers
         {
             return _context.JobItems.Any(e => e.Id == id);
         }
-
-        [HttpGet("/register/username={username}&email={email}&password={password}")]
-        public bool GetVerify(string username, string email, string password)
-        {
-            foreach (var user in _userContext.Users)
-            {
-                if (user.Email == email || user.Username == username)
-                {
-                    return false;
-                }
-            }
-            Save(username,email,password);
-            return true;
-
-        }
-
-        [HttpGet("/getCookieData")]
-        public string GetCookieData()
-        {
-            var user = HttpContext.User;
-            return user.Identity.Name;
-        }
-
-        public void Save(string username, string email, string password)
-        {
-            var hashedPassword = _hashService.Hash(password);
-            _userContext.Users.Add(new User(username, email, hashedPassword));
-            _userContext.SaveChanges();
-        }
-
-        [HttpGet("/login/username={username}&password={password}")]
-        public bool Login(string username, string password)
-        {
-            var users = _userContext.Users.ToList();
-            var loginService = new LoginService(username, password, users);
-            
-            if (loginService.Login())
-            {
-                CreateCookie(username);
-                return true;
-            }else
-            {
-                return false;
-            }
-        }
-
-        [HttpGet("/logout")]
-        public RedirectResult Logout(string username, string password)
-        {
-            HttpContext.SignOutAsync();
-            return Redirect("/");
-        }
-
-        private async void CreateCookie(string username)
-        {
-            var claims = new List<Claim>
-                {
-                    new Claim(ClaimTypes.Name,username)
-                };
-            var identity = new ClaimsIdentity(
-                claims, CookieAuthenticationDefaults.AuthenticationScheme
-                );
-            var principal = new ClaimsPrincipal(identity);
-            var props = new AuthenticationProperties();
-            await HttpContext.SignInAsync(CookieAuthenticationDefaults.AuthenticationScheme, principal, props);
-        }
-
     }
-
-   
 }
